requirements.txt
coverage
flake8
Flask
Flask-Login
Flask-Migrate
Flask-SQLAlchemy
Flask-WTF
Flask_Testing
Flask-Bootstrap
Flask_Mail
Flask-Talisman
psycopg2-binary
python-dotenv
itsdangerous
rncryptor
redis
celery
sentry-sdk[flask]